chore(examples): add calendar explorer demo app
