+++ examples/apps/calendar-explorer/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "calendar-explorer"
+publish = false
+license.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+
+[dependencies]
+color-eyre.workspace = true
+crossterm.workspace = true
+ratatui.workspace = true
+time = "0.3.37"
+
+[lints]
+workspace = true
