+++ ratatui/Cargo.toml
@@ -143,11 +143,6 @@ name = "main"
 harness = false
 
 [[example]]
-name = "calendar"
-required-features = ["crossterm", "widget-calendar"]
-doc-scrape-examples = true
-
-[[example]]
 name = "canvas"
 required-features = ["crossterm"]
 doc-scrape-examples = true
